Refactor prompts
