+++ src/components/Elements/Prompt.js
@@ -1,40 +1,17 @@
 import React, { Component } from 'react';
-import Touch from 'react-hammerjs';
-import Pips from './Pips';
-
-const promptClasses = (index, currentIndex) => {
-  return currentIndex === index ? 'prompt__prompt prompt__prompt--active' : 'prompt__prompt';
-}
 
 class Prompt extends Component {
   render() {
 
     const {
-      prompts,
-      currentIndex,
-      handleTap,
-      handleSwipe
+      prompt,
+      isActive
     } = this.props;
 
     return (
-      <Touch onTap={ handleTap } onSwipe={ handleSwipe } >
-        <div className='prompt'>
-          <div className='prompt__prompts'>
-            { prompts.map((prompt, index) => {
-
-              return (
-                <div key={ index } className={ promptClasses(index, currentIndex) }>
-                  <h3>{ prompt.title }</h3>
-                </div>
-              );
-            }) }
-          </div>
-
-          <div className='prompt__pips'>
-            <Pips count={ prompts.length } currentIndex={ currentIndex } />
-          </div>
-        </div>
-      </Touch>
+      <div className={ isActive ? 'prompts__prompt prompts__prompt--active' : 'prompts__prompt' }>
+        <h3>{ prompt.title }</h3>
+      </div>
     );
   }
 }
