+++ src/styles/components/_all.scss
@@ -1,7 +1,7 @@
 @import 'menu';
 @import 'node';
 @import 'nodeList';
-@import 'prompt';
+@import 'prompts';
 @import 'pips';
 @import 'interface';
 @import 'protocol';
