+++ src/styles/components/_prompts.scss
@@ -1,4 +1,4 @@
-.prompt {
+.prompts {
   &__prompts {
     box-sizing: border-box;
     width: 100%;
